# openapi.yaml
openapi: 3.0.3
info:
  title: Purrfect API
  version: 1.0.0
  description: >
    Secure cat marketplace backend with escrow milestone settlement,
    veterinary inspection window, and dispute moderation.
servers:
  - url: https://api.purrfect.kz
    description: Production
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /disputes:
    post:
      tags: [disputes]
      summary: Open dispute for an order
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDisputeRequest'
      responses:
        '201':
          description: Dispute created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [disputes]
      summary: List disputes (role-aware)
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [OPEN, UNDER_REVIEW, RESOLVED_REFUND_FULL, RESOLVED_REFUND_PARTIAL, RESOLVED_RELEASE_SELLER, REJECTED]
      responses:
        '200':
          description: Dispute page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /disputes/{disputeId}:
    get:
      tags: [disputes]
      summary: Get dispute detail
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/DisputeIdParam'
      responses:
        '200':
          description: Dispute detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dispute'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /disputes/{disputeId}/evidence:
    post:
      tags: [disputes]
      summary: Upload dispute evidence
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/DisputeIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDisputeEvidenceRequest'
      responses:
        '201':
          description: Evidence added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DisputeEvidence'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
    get:
      tags: [disputes]
      summary: List dispute evidence
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/DisputeIdParam'
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Evidence page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputeEvidence'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
  /moderation/disputes:
    get:
      tags: [moderation]
      summary: List moderation dispute queue
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
        - in: query
          name: status
          schema:
            type: string
            enum: [OPEN, UNDER_REVIEW]
          example: OPEN
        - in: query
          name: sort
          schema:
            type: string
            enum: [-openedAt, priority]
          example: -openedAt
      responses:
        '200':
          description: Moderation queue page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDisputes'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /disputes/{disputeId}/reopen:
    post:
      tags: [disputes]
      summary: Reopen previously resolved dispute
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/DisputeIdParam'
      responses:
        '200': { description: Dispute reopened }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /disputes/{disputeId}/comments:
    get:
      tags: [disputes]
      summary: List comments on a dispute
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: disputeId
          required: true
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Comments page
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string, format: uuid }
                        disputeId: { type: string, format: uuid }
                        authorId: { type: string, format: uuid }
                        text: { type: string }
                        createdAt: { type: string, format: date-time }
                  meta:
                    type: object
                    properties:
                      hasNext: { type: boolean }
                      nextCursor: { type: string, nullable: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [disputes]
      summary: Post a comment to a dispute thread
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: disputeId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '201':
          description: Comment created
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/UnprocessableEntity' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CursorParam:
      in: query
      name: cursor
      required: false
      schema:
        type: string
      description: Opaque keyset cursor from previous page
      example: eyJjcmVhdGVkQXQiOiIyMDI2LTA0LTA3VDEzOjQyOjAwWiIsImlkIjoiY2YxMiJ9
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    DisputeIdParam:
      in: path
      name: disputeId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: BAD_REQUEST
            message: Invalid query parameter value
    Unauthorized:
      description: Authentication failed or token missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Access token is invalid or expired
    Forbidden:
      description: Authenticated but missing required role/permission
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Admin role is required
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: Listing not found
    Conflict:
      description: Business state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CONFLICT
            message: Order is already in disputed state
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: VALIDATION_ERROR
            message: inspection report URL is required
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: INTERNAL_ERROR
            message: Unexpected server error
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
...
    CreateDisputeRequest:
      type: object
      required: [orderId, reasonCode]
      properties:
        orderId:
          type: string
          format: uuid
        reasonCode:
          type: string
          enum: [HEALTH_MISMATCH, FAKE_BREED_DOCS, NOT_AS_DESCRIBED, OTHER]
          example: HEALTH_MISMATCH
        evidenceUrls:
          type: array
          items:
            type: string
            format: uri
          example:
            - https://files.purrfect.kz/evidence/photo1.jpg
            - https://files.purrfect.kz/evidence/vet-note.pdf
        description:
          type: string
          example: Vet found congenital issue not disclosed in listing.
...
    Dispute:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, RESOLVED_REFUND_FULL, RESOLVED_REFUND_PARTIAL, RESOLVED_RELEASE_SELLER, REJECTED]
        reasonCode:
          type: string
        openedAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
          nullable: true
    AddDisputeEvidenceRequest:
      type: object
      required: [evidenceType, fileUrl]
      properties:
        evidenceType:
          type: string
          enum: [PHOTO, VIDEO, PDF_REPORT, CHAT_SCREENSHOT, OTHER]
          example: PDF_REPORT
        fileUrl:
          type: string
          format: uri
          example: https://files.purrfect.kz/evidence/report_224.pdf
        note:
          type: string
          example: Vet report confirming congenital heart issue.
    DisputeEvidence:
      type: object
      properties:
        id:
          type: string
          format: uuid
        disputeId:
          type: string
          format: uuid
        uploaderId:
          type: string
          format: uuid
        evidenceType:
          type: string
          enum: [PHOTO, VIDEO, PDF_REPORT, CHAT_SCREENSHOT, OTHER]
        fileUrl:
          type: string
          format: uri
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    PageMeta:
      type: object
      required: [hasNext, nextCursor]
      properties:
        hasNext:
          type: boolean
        nextCursor:
          type: string
          nullable: true
...
    PaginatedDisputes:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dispute'
        meta:
          $ref: '#/components/schemas/PageMeta'
    PaginatedDisputeEvidence:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/DisputeEvidence'
        meta:
          $ref: '#/components/schemas/PageMeta'
...